\docType{data}
\name{gauche_hidden_data}
\alias{gauche_hidden_data}
\title{gauche_hidden_data}
\format{A \code{data.frame} with 5 mutations (rows) measured at 3 time points (columns)}
\usage{
data(gauche_hidden_data)
}
\description{
A synthetic data set provided to illustrate how to use
gauche.  See the accompanying vignette for more details.
}
\author{
Alex Murison \email{Alexander.Murison@icr.ac.uk} and
Christopher Wardell \email{Christopher.Wardell@icr.ac.uk}
}
\seealso{
\code{\link{ga-class}}, \code{\link{ga}},
\code{\link{gaucheReport}},
\code{\link{gauche_simple_data}},
\code{\link{gauche_hidden_data}},
\code{\link{gauche_synth_data}},
\code{\link{gauche_synth_data_jittered}},
\code{\link{BYB1_G07_pruned}}
}

