\docType{data}
\name{gauche_synth_data_jittered}
\alias{gauche_synth_data_jittered}
\title{gauche_synth_data_jittered}
\format{A \code{data.frame} with 90 genes (rows) measured at 4 time points (columns), with added noise.}
\usage{
data(gauche_synth_data_jittered)
}
\description{
A synth data set provided to illustrate how to use gauche.
See the accompanying vignette for more details.
}
\author{
Alex Murison \email{Alexander.Murison@icr.ac.uk} and
Christopher Wardell \email{Christopher.Wardell@icr.ac.uk}
}
\seealso{
\code{\link{ga-class}}, \code{\link{ga}},
\code{\link{gaucheReport}},
\code{\link{gauche_simple_data}},
\code{\link{gauche_hidden_data}},
\code{\link{gauche_synth_data}},
\code{\link{gauche_synth_data_jittered}},
\code{\link{BYB1_G07_pruned}}
}

